We're leaving together, But still it's farewell And maybe we'll come back To earth, who can tell? I guess there is no one to blame We're leaving ground (leaving ground) Will things ever be the same again?
It's the final countdown
The final countdown

Oh

We're heading for Venus (Venus)
And still we stand tall
'Cause maybe they've seen us (seen us)
And welcome us all, yeah
With so many light years to go
And things to be found (to be found)
I'm sure that we'll all miss her so

It's the final countdown
The final countdown
The final countdown
The final countdown
Oh

The final countdown,oh
It's the final count down
The final countdown
The final countdown
The final countdown
Oh
It's the final count down
We're leaving together
The final count down
We'll all miss her so
It's the final countdown
It's the final countdown
Oh
It's the final countdown, yeah

+++ src/test/java/utwente/ns/ip/SimulatedHRP4Layer.java
@@ -54,7 +54,7 @@ public class SimulatedHRP4Layer implements IHRP4Layer {
     public synchronized void addReceiveListener(IReceiveListener receiver) {
         this.receiveListeners.add(receiver);
         if (receiver instanceof IHRP4Socket) {
-        	ports.add(((IHRP4Socket) receiver).getDstPort());
+        	ports.add((int) ((IHRP4Socket) receiver).getDstPort());
 		}
     }
